fix: apply types separation for browser instrumentations
