+++ plugins/web/opentelemetry-instrumentation-document-load/src/index.ts
@@ -15,3 +15,4 @@
  */
 
 export * from './instrumentation';
+export * from './enums/AttributeNames';
