+++ plugins/web/opentelemetry-instrumentation-long-task/src/index.ts
@@ -15,3 +15,4 @@
  */
 
 export * from './instrumentation';
+export * from './types';
