+++ plugins/web/opentelemetry-instrumentation-user-interaction/src/index.ts
@@ -15,3 +15,5 @@
  */
 
 export * from './instrumentation';
+export * from './types';
+export * from './enums/AttributeNames';
