+++ plugins/web/opentelemetry-plugin-react-load/src/index.ts
@@ -15,3 +15,4 @@
  */
 
 export * from './BaseOpenTelemetryComponent';
+export * from './enums/AttributeNames';
