+++ plugins/web/opentelemetry-plugin-react-load/src/internal-types.ts
